refactor: 주문관련 vuex 로직 분리
